backend/requirements.txt
fastapi
uvicorn[standard]
pillow
python-multipart
sentence-transformers
faiss-cpu
numpy
llama-cpp-python
requests


